requirements.txt
fastapi[standard]>=0.115.5
motor>=3.6.0
pydantic-settings>=2.6.1
pydantic>=2.9.2
pyyaml>=6.0.2
slowapi>=0.1.9
tiktoken>=0.8.0
asyncstdlib>=3.13.0
ujson>=5.10.0
asgiref>=3.8.1
curl_cffi
